Fix some of the codex recommendations
